Add option to disable linking of libcxx into fuzzer
Summary:
The ability to partially link libcxx into fuzzer does not work for cross
compilations.  CMAKE_EXE_LINKER_FLAGS, etc are not passed to the custom CMake
invocation and cross-compiling libcxx without a libcxxabi, as is done here, has
further issues as well.

Add an option to disable linking of libcxx into fuzzer.  The default for this
option is to link libcxx.

Reviewers: srhines, chh, phosek, EricWF

Subscribers: mgorny, delcypher, #sanitizers, llvm-commits

Differential Revision: https://reviews.llvm.org/D44532

Change-Id: I013e5450ff7d1c900505b45e0b02c3f2ff10f571