[builtins] Rounding mode support for addxf3/subxf3
Implement rounding mode support for addxf3/subxf3.

On architectures that implemented the support, this will access the
corresponding floating point environment register to apply the correct
rounding. For other architectures, it will keep the current behaviour
and use IEEE-754 default rounding mode (to nearest, ties to even).

ARM32/AArch64 support implemented in this change. i386 and AMD64 will
be added in a follow up change.

Change-Id: Idd8044ffb121e82d1ec2160b11cfcf5b5805167d